Refactored addon tree.
