+++ src/addon/main.h
@@ -1,10 +1,10 @@
 #ifndef ___NODE_SYPHON_H___
 #define ___NODE_SYPHON_H___
 
-#include "server/OpenGLServer.h"
-#include "server/MetalServer.h"
-#include "client/OpenGLClient.h"
-#include "client/MetalClient.h"
+#include "opengl/OpenGLServer.h"
+#include "opengl/OpenGLClient.h"
+#include "metal/MetalServer.h"
+#include "metal/MetalClient.h"
 #include "directory/ServerDirectory.h"
 
 #endif
