+++ src/addon/opengl/OpenGLServer.mm
@@ -2,9 +2,10 @@
 #include <iostream>
 
 #include "OpenGLServer.h"
+#include "OpenGLHelper.h"
 
 #import "../helpers/ServerDescriptionHelper.h"
-#include "../helpers/OpenGLHelper.h"
+
 
 using namespace syphon;
